+++ src/olmo_core/utils.py
@@ -23,6 +23,7 @@ from rich.console import Console, ConsoleRenderable
 from rich.highlighter import NullHighlighter
 from rich.text import Text
 from rich.traceback import Traceback
+from torch.utils.flop_counter import FlopCounterMode
 
 from .config import StrEnum
 from .exceptions import OLMoCLIError, OLMoEnvironmentError, OLMoError, OLMoThreadError
@@ -149,6 +150,13 @@ def get_default_device() -> torch.device:
         return torch.device("mps")
     else:
         return torch.device("cpu")
+
+
+def has_compute_capability(major: int, minor: int) -> bool:
+    return torch.cuda.is_available() and torch.cuda.get_device_capability() >= (
+        major,
+        minor,
+    )
 
 
 def seed_all(seed: int):
@@ -629,17 +637,38 @@ def capped_powers_of_2(x: int, cap: int) -> List[int]:
 
 
 def format_float(value: float) -> str:
-    if value == 0.0:
+    if math.isnan(value):
+        return "nan"
+    elif math.isinf(value):
+        return "inf" if value > 0 else "-inf"
+    abs_value = abs(value)
+    if abs_value == 0.0:
         return "0.0"
-    elif value < 0.0001:
+    elif abs_value >= 1e9:
+        for suffix, factor in (("E", 1e18), ("P", 1e15), ("T", 1e12), ("B", 1e9)):
+            if abs_value >= factor:
+                scaled = value / factor
+                abs_scaled = abs(scaled)
+                if abs_scaled > 100:
+                    decimals = 1
+                elif abs_scaled > 10:
+                    decimals = 2
+                elif abs_scaled > 1:
+                    decimals = 3
+                else:
+                    decimals = 4
+                scaled_str = f"{scaled:,.{decimals}f}"
+                return f"{scaled_str}{suffix}"
+
+    if abs_value < 0.0001:
         return f"{value:.2E}"
-    elif value > 1000:
+    elif abs_value >= 1000:
         return f"{int(value):,d}"
-    elif value > 100:
+    elif abs_value > 100:
         return f"{value:.1f}"
-    elif value > 10:
+    elif abs_value > 10:
         return f"{value:.2f}"
-    elif value > 1:
+    elif abs_value > 1:
         return f"{value:.3f}"
     else:
         return f"{value:.4f}"
@@ -758,3 +787,32 @@ def get_or_init_stream(id: str, priority: int = 0) -> torch.cuda.Stream:
         stream = cast(torch.cuda.Stream, torch.cuda.Stream(priority=priority))
         _CUDA_STREAMS[id] = stream
         return stream
+
+
+def record_flops(
+    model: torch.nn.Module,
+    inp: Union[torch.Tensor, tuple],
+    with_backward: bool = False,
+    display: bool = False,
+) -> int:
+    # Source: https://alessiodevoto.github.io/Compute-Flops-with-Pytorch-built-in-flops-counter/
+    istrain = model.training
+    model.eval()
+
+    inp = inp if isinstance(inp, torch.Tensor) else torch.randn(inp)
+
+    # FlopCounterMode has some limitations, for example if activation recomputation is used,
+    # it will count the flops for the recomputation as well. This applies to SDPA kernels as well,
+    # since they use recomputation internally (a la flash attention). Additionally, if custom kernels
+    # are used, it will not count the flops for them unless a custom flop formula is registered.
+    # https://github.com/pytorch/pytorch/issues/123800
+    flop_counter = FlopCounterMode(display=display, depth=999999)
+    with flop_counter:
+        if with_backward:
+            model(inp).sum().backward()
+        else:
+            model(inp)
+    total_flops = flop_counter.get_total_flops()
+    if istrain:
+        model.train()
+    return total_flops